refactor(qmatmul): extract QLinear/QGgufVarBuilder and wire Qwen3.5
Move QLinear, QGgufVarBuilder and qlinear_b from gemma4.rs into a new
shared quantized_linear module, then thread QGgufVarBuilder through the
Qwen3.5 model hierarchy (Qwen35Model → DecoderLayer → FullAttention /
LinearAttn) in preparation for replacing linear_no_bias calls with
qlinear_b to enable the quantized GEMV path for decode.
