+++ inferrs/src/models/gemma4.rs
@@ -20,11 +20,11 @@
 //!   `embed_tokens_per_layer` is scaled by `sqrt(hidden_size_per_layer_input)`.
 //! * All language-model weights live under `model.language_model.*`.
 
-use candle_core::quantized::QMatMul;
 use candle_core::{DType, Device, Module, Result, Tensor, D};
 use candle_nn::{rms_norm, Activation, RmsNorm, VarBuilder};
 use std::sync::Arc;
 
+use crate::models::quantized_linear::{qlinear_b, QGgufVarBuilder, QLinear};
 use crate::turbo_quant::{TurboQuantConfig, TurboQuantKvCache, MIN_KV_BUFFER_CAP};
 
 // ---------------------------------------------------------------------------
@@ -238,353 +238,6 @@ impl PliEmbeddingTable {
     ) -> candle_core::Result<candle_core::Tensor> {
         self.lookup_batch(&[token_id], embed_dim, dtype)
             .and_then(|t| t.reshape((1usize, 1usize, embed_dim)))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// QLinear: a Linear layer backed by either a standard Tensor or a QMatMul.
-//
-// When weights are loaded from a GGUF file with --quantize, using QMatMul keeps
-// the weights in their compressed quantized form (e.g. Q4K) and dispatches to
-// Metal's optimised quantized GEMV kernel (call_quantized_matmul_mv_t) during
-// the decode step.  This is the same kernel that llama.cpp/ggml uses and gives
-// ~3-4× higher decode throughput compared to dequantizing to bf16 first.
-//
-// For safetensors (bf16) models, QLinear falls back to a standard Linear
-// (identical to the previous behaviour).
-// ---------------------------------------------------------------------------
-
-/// A linear projection layer backed by `QMatMul`.
-///
-/// ## Memory-efficient quantized linear (GGUF path)
-///
-/// Stores only the `QMatMul::QTensor` — the weight stays compressed in Metal
-/// memory (Q4K ≈ 4.5 bits/param).  No second bf16 copy is kept.
-///
-/// * **Decode** (seq_len = 1): Metal's `kernel_mul_mv_q4_K_f32` GEMV.
-///   Input is cast bf16→f32 (the kernel requires f32), output cast back.
-///   Q4K is 4× smaller than bf16 → ~3-4× faster GEMV.
-///
-/// * **Prefill** (seq_len > 1): `forward_via_f16` dequantizes the QTensor to
-///   f16 on-the-fly, runs the standard f16 GEMM, then converts back to the
-///   original dtype.  The dequantization is a single fast Metal kernel; its
-///   cost is negligible compared with the GEMM for any realistic sequence length.
-///   Memory stays at QTensor-only — no permanent second copy.
-///
-/// ## Safetensors path
-///
-/// `inner` is `QMatMul::Tensor` (plain bf16).  Both paths use the standard
-/// matmul, identical to `candle_nn::Linear`.
-#[derive(Debug, Clone)]
-pub struct QLinear {
-    inner: QMatMul,
-    pub(crate) bias: Option<Tensor>,
-}
-
-impl QLinear {
-    /// Build from a quantized tensor (GGUF path).
-    pub fn from_qtensor(
-        qtensor: Arc<candle_core::quantized::QTensor>,
-        bias: Option<Tensor>,
-    ) -> Result<Self> {
-        let inner = QMatMul::from_arc(qtensor)?;
-        Ok(Self { inner, bias })
-    }
-
-    /// Build from a regular tensor (safetensors path).
-    pub fn from_tensor(weight: Tensor, bias: Option<Tensor>) -> Self {
-        Self {
-            inner: QMatMul::Tensor(weight),
-            bias,
-        }
-    }
-
-    /// Returns true when the underlying weight is a quantized QTensor (GGUF path).
-    /// Returns false for the dense BF16 safetensors path (`QMatMul::Tensor`).
-    pub fn is_quantized(&self) -> bool {
-        matches!(self.inner, QMatMul::QTensor(_))
-    }
-}
-
-impl Module for QLinear {
-    fn forward(&self, xs: &Tensor) -> Result<Tensor> {
-        match &self.inner {
-            QMatMul::QTensor(_) => {
-                let orig_dtype = xs.dtype();
-                // On CUDA with BF16 activations, the patched `dequantize_matmul_vec`
-                // has a BF16 fast path that fuses BF16→Q8_1 in one kernel dispatch
-                // (vs the old two-dispatch BF16→F32 + F32→Q8_1 path).
-                // Pass BF16 input directly and save one kernel launch per GEMV.
-                // Non-CUDA or non-BF16: keep the standard F32 conversion path.
-                let r = if matches!(xs.device(), candle_core::Device::Cuda(_))
-                    && orig_dtype == DType::BF16
-                {
-                    // Direct BF16 path: skip the BF16→F32 conversion kernel.
-                    self.inner.forward(xs)?
-                } else {
-                    let xs_f32 = if orig_dtype == DType::F32 {
-                        xs.clone()
-                    } else {
-                        xs.to_dtype(DType::F32)?
-                    };
-                    self.inner.forward(&xs_f32)?
-                };
-                // GEMV output is always F32; convert back to orig_dtype if needed.
-                let result = if orig_dtype == DType::F32 || r.dtype() == orig_dtype {
-                    r
-                } else {
-                    r.to_dtype(orig_dtype)?
-                };
-                match &self.bias {
-                    None => Ok(result),
-                    Some(b) => result.broadcast_add(b),
-                }
-            }
-            _ => {
-                // Dense path (safetensors bf16): standard matmul.
-                let result = self.inner.forward(xs)?;
-                match &self.bias {
-                    None => Ok(result),
-                    Some(b) => result.broadcast_add(b),
-                }
-            }
-        }
-    }
-}
-
-impl QLinear {
-    /// Forward pass that takes an already-F32 input and returns F32 output.
-    ///
-    /// When the underlying QMatMul is a QTensor (quantized GGUF path), the
-    /// CUDA/Metal GEMV kernel requires F32 input.  If the caller has already
-    /// converted the activation to F32 (e.g. to amortise the cost across
-    /// multiple QLinear calls that share the same input), calling this method
-    /// directly skips the two dtype-conversion kernel launches that
-    /// `forward()` would add.
-    ///
-    /// For the Dense path (QMatMul::Tensor), this falls through to a
-    /// standard matmul and then converts the result to F32 if needed.
-    #[allow(dead_code)]
-    pub fn forward_f32(&self, xs_f32: &Tensor) -> Result<Tensor> {
-        debug_assert_eq!(xs_f32.dtype(), DType::F32, "forward_f32 requires F32 input");
-        match &self.inner {
-            QMatMul::QTensor(_) => {
-                // No conversion needed — the GEMV kernel already takes F32.
-                let r = self.inner.forward(xs_f32)?;
-                // r is F32; return F32 to the caller.
-                match &self.bias {
-                    None => Ok(r),
-                    Some(b) => {
-                        let b_f32 = if b.dtype() == DType::F32 {
-                            b.clone()
-                        } else {
-                            b.to_dtype(DType::F32)?
-                        };
-                        r.broadcast_add(&b_f32)
-                    }
-                }
-            }
-            _ => {
-                // Dense path: standard matmul, result may be bf16; cast to F32.
-                let result = self.inner.forward(xs_f32)?;
-                let result = if result.dtype() == DType::F32 {
-                    result
-                } else {
-                    result.to_dtype(DType::F32)?
-                };
-                match &self.bias {
-                    None => Ok(result),
-                    Some(b) => {
-                        let b_f32 = b.to_dtype(DType::F32)?;
-                        result.broadcast_add(&b_f32)
-                    }
-                }
-            }
-        }
-    }
-}
-
-// NVFP4 dequantization is implemented in crate::nvfp4 (shared with quantize.rs).
-
-// ---------------------------------------------------------------------------
-// QGgufVarBuilder: a VarBuilder for quantized GGUF tensors.
-//
-// Provides Arc<QTensor> access for named tensors, enabling QMatMul::QTensor
-// construction that uses the Metal quantized GEMV kernel path.
-// ---------------------------------------------------------------------------
-
-/// A VarBuilder that holds tensors from a GGUF file in their original
-/// quantized form (rather than dequantizing them upfront to bf16).
-///
-/// Used alongside the standard `VarBuilder` (for non-quantized tensors such
-/// as norms and embeddings) to provide `QLinear`-backed projections.
-#[derive(Clone)]
-pub struct QGgufVarBuilder {
-    data: Arc<std::collections::HashMap<String, Arc<candle_core::quantized::QTensor>>>,
-    path: Vec<String>,
-}
-
-impl QGgufVarBuilder {
-    /// Load projection tensors from a GGUF file, retaining their quantized format.
-    ///
-    /// Only tensors that are used as linear projection weights (i.e. those that
-    /// will be called via `QLinear` for GEMV/GEMM) are loaded into Metal memory
-    /// here.  Embedding tables (`embed_tokens`, `embed_tokens_per_layer`) are
-    /// deliberately excluded: they are used only for index-select lookups and
-    /// do not benefit from the quantized GEMV kernel path.  They are loaded
-    /// separately via the standard `VarBuilder` (dequantized to bf16) and can
-    /// be very large (e.g. `embed_tokens_per_layer` is 4.7 GB in bf16 for E2B).
-    /// Loading them twice would double their memory footprint.
-    ///
-    /// Exception: `embed_tokens.weight` (the tied lm_head) IS loaded here
-    /// because it is also used as the output projection (lm_head GEMV), which
-    /// is the single most expensive per-token operation.
-    pub fn from_gguf<P: AsRef<std::path::Path>>(
-        p: P,
-        device: &Device,
-    ) -> candle_core::Result<Self> {
-        use candle_core::quantized::gguf_file;
-        let mut file = std::fs::File::open(p.as_ref()).map_err(candle_core::Error::from)?;
-        let content = gguf_file::Content::read(&mut file)?;
-        let mut data = std::collections::HashMap::new();
-        for tensor_name in content.tensor_infos.keys() {
-            // Skip the PLI embedding table — it is loaded via the memory-efficient
-            // GGUF file-backed PliEmbeddingTable::GgufFile which reads rows on
-            // demand without loading 1.9 GB into CPU RAM.
-            if tensor_name.contains("embed_tokens_per_layer") {
-                continue;
-            }
-            let qt = content.tensor(&mut file, tensor_name, device)?;
-            data.insert(tensor_name.to_string(), Arc::new(qt));
-        }
-        Ok(Self {
-            data: Arc::new(data),
-            path: Vec::new(),
-        })
-    }
-
-    /// Enter a sub-namespace (mirrors `VarBuilder::pp`).
-    pub fn pp<S: ToString>(&self, s: S) -> Self {
-        let mut path = self.path.clone();
-        path.push(s.to_string());
-        Self {
-            data: self.data.clone(),
-            path,
-        }
-    }
-
-    /// Build the fully-qualified name for a tensor under the current namespace.
-    pub fn full_name(&self, name: &str) -> String {
-        if self.path.is_empty() {
-            name.to_string()
-        } else {
-            format!("{}.{}", self.path.join("."), name)
-        }
-    }
-
-    /// Retrieve the raw `Arc<QTensor>` for the "weight" tensor at the current path.
-    ///
-    /// Returns `None` if the tensor is not found in the GGUF data map.
-    pub fn get_qtensor(&self) -> Option<Arc<candle_core::quantized::QTensor>> {
-        let name = self.full_name("weight");
-        self.data.get(&name).cloned()
-    }
-
-    /// Build a bias-free `QLinear` for the "weight" tensor at the current path.
-    ///
-    /// Returns `QMatMul::QTensor` (fast Metal quantized GEMV) when the tensor
-    /// is present; errors if absent.
-    /// Build a bias-free `QLinear` from the "weight" tensor at the current path.
-    pub fn qlinear_weight(&self) -> Result<QLinear> {
-        let name = self.full_name("weight");
-        match self.data.get(&name) {
-            Some(qt) => QLinear::from_qtensor(qt.clone(), None),
-            None => candle_core::bail!("QGgufVarBuilder: tensor not found: {name}"),
-        }
-    }
-
-    /// Try to build a `QLinear`; returns `None` if the tensor is absent.
-    pub fn try_qlinear_weight(&self) -> Option<Result<QLinear>> {
-        let name = self.full_name("weight");
-        self.data
-            .get(&name)
-            .map(|qt| QLinear::from_qtensor(qt.clone(), None))
-    }
-}
-
-/// Build a bias-free QLinear layer.
-///
-/// If `qvb` is `Some`, keeps the weight as QTensor (quantized GGUF path).
-/// If `qvb` is `None`, loads the weight from safetensors and online-quantizes
-/// it to Q8_0 so that the Metal/CUDA quantized GEMV kernel is used at decode
-/// time instead of a full BF16 GEMM.  Q8_0 is near-lossless (8-bit absmax
-/// per 32-element block) and halves the memory bandwidth required for each
-/// decode GEMV compared with BF16, giving ~2–3× decode throughput improvement.
-///
-/// NVFP4 weights are dequantized first (to BF16) and then re-quantized to Q8_0.
-///
-/// Both `vb` and `qvb` are already `.pp("layer_name")` scoped.
-fn qlinear_b(
-    in_dim: usize,
-    out_dim: usize,
-    bias: bool,
-    vb: VarBuilder,
-    qvb: Option<&QGgufVarBuilder>,
-) -> Result<QLinear> {
-    // Load bias from the dense VarBuilder when requested.  The GGUF path also
-    // uses vb for bias since bias vectors are stored at F16 (not quantized).
-    let b = if bias {
-        Some(vb.get(out_dim, "bias")?)
-    } else {
-        None
-    };
-    if let Some(q) = qvb {
-        let mut ql = q.qlinear_weight()?;
-        ql.bias = b;
-        Ok(ql)
-    } else {
-        // Check for NVFP4 quantized weight (U8 packed FP4 + F8E4M3 block scales).
-        let dtype = vb.dtype();
-        let device = vb.device().clone();
-        let weight = if let Some(w) =
-            crate::nvfp4::try_load_from_varbuilder(&vb, out_dim, in_dim, dtype, &device)?
-        {
-            w
-        } else {
-            vb.get((out_dim, in_dim), "weight")?
-        };
-
-        // Online Q4K quantization: convert BF16 weights to a QTensor so that
-        // Metal's quantized GEMV kernel (call_quantized_matmul_mv_t) is used at
-        // decode time instead of a full dense matmul.  Q4K (Q4_K_M) uses 256-
-        // element blocks; all major Gemma4 projection dimensions are multiples
-        // of 256.  For any tensor whose element count is not a multiple of 256
-        // (e.g. small adapter layers) we fall back to Q8_0 (block_size=32), and
-        // if that also fails we keep the original dense BF16 path.
-        //
-        // Q4K is ~4.5 bits/weight — 4× less bandwidth than BF16 GEMV — matching
-        // the decode throughput of a pre-built Q4K GGUF without requiring the
-        // user to pass --quantize.
-        let elem_count = weight.elem_count();
-        let quant_dtype = if elem_count % 256 == 0 {
-            Some(candle_core::quantized::GgmlDType::Q4K)
-        } else if elem_count % 32 == 0 {
-            Some(candle_core::quantized::GgmlDType::Q8_0)
-        } else {
-            None
-        };
-        if let Some(dtype) = quant_dtype {
-            match candle_core::quantized::QTensor::quantize(&weight, dtype) {
-                Ok(qt) => return QLinear::from_qtensor(Arc::new(qt), b),
-                Err(e) => {
-                    tracing::debug!(
-                        "online {dtype:?} quantization failed for [{out_dim}×{in_dim}], \
-                         falling back to BF16: {e}"
-                    );
-                }
-            }
-        }
-        Ok(QLinear::from_tensor(weight, b))
     }
 }
